test/requirements.txt
Flask==2.1.0
requests==2.26.0
pymysql==1.0.2
python-dotenv==0.19.2
charset==2.0.12
click==8.2.1
colorama==0.4.6
flask-cors==6.0.0
PyMySQL==1.0.2
python==dotenv